+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Fixed `ValueDecoration` when all items were 0
 * Add value label generator in `ValueDecoration`
 * Add show the line for value to `HorizontalDecoration` for more precise control of when to show horizontal line
+* Fix fixed horizontal decoration values getting clipped
 
 ## [2.0.0+1] - 01.03.2022.
 * Updated packages