+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [2.0.0+2] - 
+## [2.0.0+2] - 05.04.2022.
 * Reduce final .apk app size when using this lib
 * Fixed `ValueDecoration` when all items were 0
 * Add value label generator in `ValueDecoration`